updated CompanyPart model to use DecimalField instead of FloatField

+++ stockupapi/models/company_parts.py
@@ -1,5 +1,5 @@
 from django.db.models.deletion import CASCADE, DO_NOTHING
-from django.db.models.fields import IntegerField, FloatField
+from django.db.models.fields import DecimalField, IntegerField, FloatField
 from django.db.models.fields.related import ForeignKey
 from safedelete.models import SafeDeleteModel
 from safedelete.models import SOFT_DELETE
@@ -10,4 +10,4 @@ class CompanyPart(SafeDeleteModel):
     part = ForeignKey("Part", on_delete=DO_NOTHING)
     in_inventory = IntegerField()
     min_required = IntegerField()
-    cost = FloatField()
+    cost = DecimalField(max_digits=25, decimal_places=2)
